requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.14.0
tensorboard>=2.13.0
numpy>=1.24.0
pillow>=9.5.0
tqdm>=4.65.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.25.0
pandas>=2.0.0
